update lcd refresh cycle hooks to update raw memory

+++ src/mem/io/graphics.rs
@@ -71,8 +71,8 @@ impl LCD {
         }
     }
 
-    pub fn update_vcount(&mut self, vcount: u32) {
-        self.vcount = vcount as u8;
+    pub fn update_vcount(&mut self, vcount: u8) {
+        self.vcount = vcount;
         self.disp_stat.vcount_triggered =
             self.vcount == self.disp_stat.vcount_line_trigger;
     }
@@ -312,7 +312,7 @@ impl DispCnt {
 pub struct DispStat {
     /// 0   (W) = V Refresh status. This will be 0 during VDraw, and 1 during VBlank.
     pub is_vblank: bool,
-    /// 1   (G) = H Refresh status. This will be 0 during HDraw, and 1 during HBlank/VBlank
+    /// 1   (G) = H Refresh status. This will be 0 during HDraw, and 1 during HBlank.
     pub is_hblank: bool,
     /// 2   (Z) = VCount Triggered Status. Gets set to 1 when a Y trigger interrupt occurs.
     pub vcount_triggered: bool,
